Changed data types for efficiency

+++ Main/Main/Leg.cpp
@@ -15,7 +15,7 @@ Leg::Leg(){
     tibiaOffset = 0;
 }
 
-Leg::Leg(int legNum, int coxaPin, int femurPin, int tibiaPin) : legNum(legNum), coxaPin(coxaPin), femurPin(femurPin), tibiaPin(tibiaPin) {
+Leg::Leg(byte legNum, byte coxaPin, byte femurPin, byte tibiaPin) : legNum(legNum), coxaPin(coxaPin), femurPin(femurPin), tibiaPin(tibiaPin) {
     pwm = Adafruit_PWMServoDriver(0x40);
     coxaLenMm = 27;
     femurLenMm = 55;
@@ -24,11 +24,11 @@ Leg::Leg(int legNum, int coxaPin, int femurPin, int tibiaPin) : legNum(legNum), 
     femurOffset = 0;
     tibiaOffset = 0;
 }
-void Leg::setAngles(double x, double y, double z){
+void Leg::setAngles(float x, float y, float z){
     
     //Finding Hypothenuse and Angle of Coxa ==========================
-    double w = sqrt(pow(x, 2) + pow(y, 2));
-    double coxaAngleR = atan2(y,x);
+    float w = sqrt(pow(x, 2) + pow(y, 2));
+    float coxaAngleR = atan2(y,x);
     //Rec Prev
     prevCoxaAngle = coxaAngle;
     coxaAngle = coxaAngleR * 57296 / 1000;
@@ -96,17 +96,17 @@ void Leg::setAngles(double x, double y, double z){
     //Finding Angle of Femur & Tibia ===================================
     
     //Calculating Distance
-    double newX = w - coxaLenMm;
+    float newX = w - coxaLenMm;
 
     //Calculating Tibia &  Femur Angle
       //Finding Distance from Point
-      double hyp = sqrt(pow(newX, 2) + pow(z, 2));
+      float hyp = sqrt(pow(newX, 2) + pow(z, 2));
       // Serial.println("");
       // Serial.println("Hypothenus");
       // Serial.println(hyp);
     
       //Alpha
-      double alpha = pow(femurLenMm, 2) + pow(tibiaLenMm, 2) - pow(newX, 2) - pow(z, 2);
+      float alpha = pow(femurLenMm, 2) + pow(tibiaLenMm, 2) - pow(newX, 2) - pow(z, 2);
       alpha = alpha / (2 * femurLenMm * tibiaLenMm);
       alpha = acos(alpha);
       alpha = (alpha / PI) * 180;
@@ -115,14 +115,14 @@ void Leg::setAngles(double x, double y, double z){
       // Serial.println(alpha);
     
       //Q2
-      double q2 = 180 - alpha;
+      float q2 = 180 - alpha;
       Serial.println("");
       Serial.println("Q2");
       Serial.println(q2);
     
       //Beta
       q2 = q2 * PI/180;
-      double beta = tibiaLenMm * sin(q2);
+      float beta = tibiaLenMm * sin(q2);
       beta = beta / (femurLenMm + tibiaLenMm * cos(q2));
       beta = atan(beta);
       beta = (beta / PI) * 180;
@@ -131,14 +131,14 @@ void Leg::setAngles(double x, double y, double z){
       // Serial.println(beta);
     
       //Gamma
-      double gamma = atan(z/newX);
+      float gamma = atan(z/newX);
       gamma = (gamma / PI) * 180;
       // Serial.println("");
       // Serial.println("Gamma");
       // Serial.println(gamma);
     
       //Q1
-      double q1 = gamma + beta;
+      float q1 = gamma + beta;
       q2 = q2 * 180/PI;
       Serial.println("Q1");
       Serial.println(q1);
@@ -205,7 +205,7 @@ void Leg::setAngles(double x, double y, double z){
 
 }
 
-int returnPulseWidth(double angle){
+int returnPulseWidth(float angle){
     int pulse_wide, analog_value;
     pulse_wide = map((int) angle, 0, 180, MIN_PULSE_WIDTH, MAX_PULSE_WIDTH);
     analog_value = int(float(pulse_wide) / 1000000 * FREQUENCY * 4096);
@@ -268,7 +268,7 @@ void Leg::lift(){
     }
 }
 
-void Leg::calibrate(double x, double y, double z){
+void Leg::calibrate(float x, float y, float z){
   Serial.println("Calibration ===========================================================");
   setAngles(x, y, z);
   
@@ -284,16 +284,16 @@ void Leg::calibrate(double x, double y, double z){
 
 }
 
-void Leg::setCoxaAngle(double angle){
+void Leg::setCoxaAngle(float angle){
 
 }
 
-void Leg::setPins(int coxa, int femur, int tibia){
+void Leg::setPins(byte coxa, byte femur, byte tibia){
     coxaPin = coxa;
     femurPin = femur;
     tibiaPin = tibia;
 }
 
-void Leg::setLegNum(int leg){
+void Leg::setLegNum(byte leg){
     legNum = leg;
 }
